requirements.txt
python==3.9.13
pandas==1.4.4
psycopg==2.9.3